refactor: chat interface 분리
